# frontend/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditEntry": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "detail": {
            "additionalProperties": true,
            "title": "Detail",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "target_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Type"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "at",
          "user_id",
          "action",
          "target_type",
          "target_id",
          "detail",
          "request_id"
        ],
        "title": "AuditEntry",
        "type": "object"
      },
      "AuditPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditEntry"
            },
            "title": "Items",
            "type": "array"
          },
          "next_before_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pass as before_id for the next (older) page",
            "title": "Next Before Id"
          }
        },
        "required": [
          "items",
          "next_before_id"
        ],
        "title": "AuditPage",
        "type": "object"
      },
...
      "CopilotAnswer": {
        "description": "The whole answer at once (``?stream=false``). The SSE stream sends the same pieces as\nevents: ``plan``, ``tool`` (one per call), ``token`` (one per line), ``citations``,\n``done``.",
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CopilotCitation"
            },
            "title": "Citations",
            "type": "array"
          },
          "engine": {
            "description": "rules (deterministic planner) or llm",
            "title": "Engine",
            "type": "string"
          },
          "intent": {
            "title": "Intent",
            "type": "string"
          },
          "refused": {
            "description": "No record found, outside the records, or not allowed",
            "title": "Refused",
            "type": "boolean"
          },
          "took_ms": {
            "title": "Took Ms",
            "type": "number"
          },
          "tools": {
            "items": {
              "$ref": "#/components/schemas/CopilotToolCall"
            },
            "title": "Tools",
            "type": "array"
          }
        },
        "required": [
          "answer",
          "citations",
          "intent",
          "tools",
          "refused",
          "engine",
          "took_ms"
        ],
        "title": "CopilotAnswer",
        "type": "object"
      },
      "CopilotAsk": {
        "additionalProperties": false,
        "properties": {
          "alert_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "The alert on screen ('this alert')",
            "title": "Alert Id"
          },
          "message": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "well_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "The well on screen ('this well')",
            "title": "Well Id"
          }
        },
        "required": [
          "message"
        ],
        "title": "CopilotAsk",
        "type": "object"
      },
      "CopilotCitation": {
        "properties": {
          "document_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "kind": {
            "description": "page (a report page) or record (a database row)",
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "n": {
            "description": "The [n] marker in the answer",
            "title": "N",
            "type": "integer"
          },
          "page_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page No"
          },
          "record_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Id"
          },
          "record_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "event | well | alert | ledger | risk",
            "title": "Record Type"
          },
          "span_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Span Ids",
            "type": "array"
          }
        },
        "required": [
          "n",
          "kind",
          "label",
          "document_id",
          "page_no",
          "span_ids",
          "filename",
          "record_type",
          "record_id"
        ],
        "title": "CopilotCitation",
        "type": "object"
      },
      "CopilotToolCall": {
        "properties": {
          "args": {
            "additionalProperties": true,
            "title": "Args",
            "type": "object"
          },
          "denied": {
            "title": "Denied",
            "type": "boolean"
          },
          "empty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Empty"
          },
          "facts": {
            "title": "Facts",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "args",
          "facts",
          "empty",
          "denied"
        ],
        "title": "CopilotToolCall",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "default": {},
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UserCreate": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "password": {
            "maxLength": 200,
            "minLength": 12,
            "title": "Password",
            "type": "string"
          },
          "roles": {
            "items": {
              "enum": [
                "viewer",
                "field_engineer",
                "rtmac_engineer",
                "drilling_engineer",
                "data_steward",
                "admin"
              ],
              "type": "string"
            },
            "minItems": 1,
            "title": "Roles",
            "type": "array"
          },
          "username": {
            "maxLength": 80,
            "minLength": 3,
            "pattern": "^[a-z0-9._-]+$",
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "name",
          "password",
          "roles"
        ],
        "title": "UserCreate",
        "type": "object"
      },
      "UserOut": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_login_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "roles": {
            "items": {
              "enum": [
                "viewer",
                "field_engineer",
                "rtmac_engineer",
                "drilling_engineer",
                "data_steward",
                "admin"
              ],
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "name",
          "roles",
          "active",
          "created_at",
          "last_login_at"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "UserUpdate": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "password": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 12,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "roles": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "viewer",
                    "field_engineer",
                    "rtmac_engineer",
                    "drilling_engineer",
                    "data_steward",
                    "admin"
                  ],
                  "type": "string"
                },
                "minItems": 1,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roles"
          }
        },
        "title": "UserUpdate",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "eRTMAC-NWIS offset-well knowledge and decision support (SIH PS 121)",
    "title": "SMRITI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/audit": {
      "get": {
        "operationId": "audit_log_api_v1_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "in": "query",
            "name": "action",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "in": "query",
            "name": "before_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audit log, newest first",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/copilot/chat": {
      "post": {
        "description": "Answers only from the read-only tools (search, events, offsets, risk profile, ledger,\nwell summary, alert explanation) the user's role allows. Every line cites a report page\nor a database record as ``[n]``; when nothing is found the answer says so. Audited.",
        "operationId": "copilot_chat_api_v1_copilot_chat_post",
        "parameters": [
          {
            "description": "false: one JSON answer instead of SSE",
            "in": "query",
            "name": "stream",
            "required": false,
            "schema": {
              "default": true,
              "description": "false: one JSON answer instead of SSE",
              "title": "Stream",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CopilotAsk"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CopilotAnswer"
                }
              },
              "text/event-stream": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask the copilot (SSE stream; ?stream=false for one JSON answer)",
        "tags": [
          "copilot"
        ]
      }
    },
...
    "/api/v1/users": {
      "get": {
        "operationId": "list_users_api_v1_users_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserOut"
                  },
                  "title": "Response List Users Api V1 Users Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List users",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "create_user_api_v1_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a user",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/users/{user_id}": {
      "patch": {
        "operationId": "update_user_api_v1_users__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change roles, name, password or deactivate",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}